+++ src/main/java/ru/careportal/core/data/UserRepo.java
@@ -5,6 +5,6 @@ import ru.careportal.core.db.model.User;
 
 import java.util.Optional;
 
-public interface UserRepo extends CrudRepository<User, Integer> {
+public interface UserRepo extends CrudRepository<User, Long> {
     Optional<User> findByEmail(String email);
 }
